requirements_nuitka.txt
requests
pydantic
httpx[http2]
Pillow>=10.0.0
